+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Better support for RTL-languages (and LTR-content in a RTL-page)
 - Users may now change their display name and password (only email accounts) on the new profile page
 - Highlighted code blocks can now use line wrapping and line numbers at once
+- Images, videos, and other non-text content is now wider in View Mode
 
 ### Changed